RR-423 - Refactor entity package structure
