--- src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/database/jpa/entity/StepEntity.kt
@@ -1,84 +1,0 @@
-package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.EnumType
-import jakarta.persistence.Enumerated
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.Id
-import jakarta.persistence.Table
-import jakarta.validation.constraints.NotNull
-import org.hibernate.Hibernate
-import org.hibernate.annotations.CreationTimestamp
-import org.hibernate.annotations.UpdateTimestamp
-import org.hibernate.annotations.UuidGenerator
-import org.springframework.data.annotation.CreatedBy
-import org.springframework.data.annotation.LastModifiedBy
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.Instant
-import java.util.UUID
-
-@Table(name = "step")
-@Entity
-@EntityListeners(AuditingEntityListener::class)
-class StepEntity(
-  @Id
-  @GeneratedValue
-  @UuidGenerator
-  var id: UUID? = null,
-
-  @Column(updatable = false)
-  @field:NotNull
-  var reference: UUID? = null,
-
-  @Column
-  @field:NotNull
-  var title: String? = null,
-
-  @Column
-  @Enumerated(value = EnumType.STRING)
-  @field:NotNull
-  var status: StepStatus? = null,
-
-  @Column
-  @field:NotNull
-  var sequenceNumber: Int? = null,
-
-  @Column(updatable = false)
-  @CreationTimestamp
-  var createdAt: Instant? = null,
-
-  @Column(updatable = false)
-  @CreatedBy
-  var createdBy: String? = null,
-
-  @Column
-  @UpdateTimestamp
-  var updatedAt: Instant? = null,
-
-  @Column
-  @LastModifiedBy
-  var updatedBy: String? = null,
-) {
-
-  override fun equals(other: Any?): Boolean {
-    if (this === other) return true
-    if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
-    other as StepEntity
-
-    return id != null && id == other.id
-  }
-
-  override fun hashCode(): Int = javaClass.hashCode()
-
-  override fun toString(): String {
-    return this::class.simpleName + "(id = $id, reference = $reference, title = $title)"
-  }
-}
-
-enum class StepStatus {
-  NOT_STARTED,
-  ACTIVE,
-  COMPLETE,
-}
