+++ src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/database/jpa/entity/actionplan/StepEntity.kt
@@ -1,16 +1,12 @@
-package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.entity
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.entity.actionplan
 
-import jakarta.persistence.CascadeType
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.EntityListeners
 import jakarta.persistence.EnumType
 import jakarta.persistence.Enumerated
-import jakarta.persistence.FetchType
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.Id
-import jakarta.persistence.JoinColumn
-import jakarta.persistence.OneToMany
 import jakarta.persistence.Table
 import jakarta.validation.constraints.NotNull
 import org.hibernate.Hibernate
@@ -20,17 +16,13 @@ import org.hibernate.annotations.UuidGenerator
 import org.springframework.data.annotation.CreatedBy
 import org.springframework.data.annotation.LastModifiedBy
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.DisplayNameAuditingEntityListener
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.DisplayNameAuditingEntityListener.CreatedByDisplayName
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.database.jpa.DisplayNameAuditingEntityListener.LastModifiedByDisplayName
 import java.time.Instant
-import java.time.LocalDate
 import java.util.UUID
 
-@Table(name = "goal")
+@Table(name = "step")
 @Entity
-@EntityListeners(value = [AuditingEntityListener::class, DisplayNameAuditingEntityListener::class])
-class GoalEntity(
+@EntityListeners(AuditingEntityListener::class)
+class StepEntity(
   @Id
   @GeneratedValue
   @UuidGenerator
@@ -45,59 +37,35 @@ class GoalEntity(
   var title: String? = null,
 
   @Column
-  @field:NotNull
-  var targetCompletionDate: LocalDate? = null,
-
-  @Column
   @Enumerated(value = EnumType.STRING)
   @field:NotNull
-  var status: GoalStatus? = null,
+  var status: StepStatus? = null,
 
   @Column
-  var notes: String? = null,
-
-  @OneToMany(cascade = [CascadeType.ALL], orphanRemoval = true, fetch = FetchType.LAZY)
-  @JoinColumn(name = "goal_id", nullable = false)
   @field:NotNull
-  var steps: MutableList<StepEntity>? = null,
+  var sequenceNumber: Int? = null,
 
   @Column(updatable = false)
   @CreationTimestamp
   var createdAt: Instant? = null,
-
-  @Column
-  @field:NotNull
-  var createdAtPrison: String? = null,
 
   @Column(updatable = false)
   @CreatedBy
   var createdBy: String? = null,
 
   @Column
-  @CreatedByDisplayName
-  var createdByDisplayName: String? = null,
-
-  @Column
   @UpdateTimestamp
   var updatedAt: Instant? = null,
 
   @Column
-  @field:NotNull
-  var updatedAtPrison: String? = null,
-
-  @Column
   @LastModifiedBy
   var updatedBy: String? = null,
-
-  @Column
-  @LastModifiedByDisplayName
-  var updatedByDisplayName: String? = null,
 ) {
 
   override fun equals(other: Any?): Boolean {
     if (this === other) return true
     if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
-    other as GoalEntity
+    other as StepEntity
 
     return id != null && id == other.id
   }
@@ -109,8 +77,8 @@ class GoalEntity(
   }
 }
 
-enum class GoalStatus {
+enum class StepStatus {
+  NOT_STARTED,
   ACTIVE,
-  COMPLETED,
-  ARCHIVED,
+  COMPLETE,
 }
